code cleaning
